gen_graph function in write added

+++ GridGraph/write.h
@@ -1,11 +1,9 @@
 #ifndef WRITE_H
 #define WRITE_H
 
-
 #include <stdio.h>
 
 // FUNCTIONS DECLARATIONS
-void write(FILE* file, int width, int height, double edge_weight_min, double edge_weight_max, int edge_count_min, int edge_count_max);
-
+void write(FILE* file, int width, int height, float edge_weight_min, float edge_weight_max, int edge_count_min, int edge_count_max);
 
 #endif
